Refactor util to typesdef, http, convert.
