+++ afscgap/convert.py
@@ -8,26 +8,10 @@ This file is part of afscgap released under the BSD 3-Clause License. See
 LICENSE.md.
 """
 import re
-import typing
 
 import requests
 
-OPT_FLOAT = typing.Optional[float]
-OPT_INT = typing.Optional[int]
-OPT_STR = typing.Optional[str]
-
-STR_OR_DICT = typing.Union[str, dict]
-RANGE_TUPLE = typing.Tuple[float]
-FLOAT_PARAM = typing.Optional[typing.Union[float, dict, RANGE_TUPLE]]
-INT_PARAM = typing.Optional[typing.Union[int, dict, RANGE_TUPLE]]
-STR_PARAM = typing.Optional[STR_OR_DICT]
-
-REQUESTOR = typing.Callable[[str], requests.Response]
-OPT_REQUESTOR = typing.Optional[REQUESTOR]
-
-TIMEOUT = 60 * 5  # 5 minutes
-
-ACCEPTABLE_CODES = [200]
+from typesdef import STR_PARAM
 
 DATE_REGEX = re.compile('(?P<month>\\d{2})\\/(?P<day>\\d{2})\\/' + \
     '(?P<year>\\d{4}) (?P<hours>\\d{2})\\:(?P<minutes>\\d{2})\\:' + \
@@ -37,24 +21,6 @@ ISO_8601_REGEX = re.compile('(?P<year>\\d{4})\\-(?P<month>\\d{2})\\-' + \
     '(?P<day>\\d{2})T(?P<hours>\\d{2})\\:(?P<minutes>\\d{2})\\:' + \
     '(?P<seconds>\\d{2})')
 ISO_8601_TEMPLATE = '%s-%s-%sT%s:%s:%s'
-
-
-def check_result(target: requests.Response):
-    """Assert that a result returned an acceptable status code.
-
-    Args:
-        target: The response to check.
-
-    Raises:
-        RuntimeError: Raised if the response returned indicates an issue or
-            unexpected status code.
-    """
-    if target.status_code not in ACCEPTABLE_CODES:
-        message = 'Got non-OK response from API: %d (%s)' % (
-            target.status_code,
-            target.text
-        )
-        raise RuntimeError(message)
 
 
 def convert_from_iso8601(target: STR_PARAM) -> STR_PARAM:
@@ -148,12 +114,3 @@ def is_iso8601(target: str) -> bool:
         True if it matches the expected format and false otherwise.
     """
     return ISO_8601_REGEX.match(target) is not None
-
-
-def build_requestor() -> REQUESTOR:
-    """Build a requestor strategy that uses the requests library.
-
-    Returns:
-        Newly built strategy.
-    """
-    return lambda x: requests.get(x, timeout=TIMEOUT)
