+++ afscgap/model.py
@@ -7,8 +7,8 @@ for Data Science and the Environment at UC Berkeley.
 This file is part of afscgap released under the BSD 3-Clause License. See
 LICENSE.md.
 """
-from afscgap.util import OPT_FLOAT
-from afscgap.util import OPT_INT
+from afscgap.typesdef import OPT_FLOAT
+from afscgap.typesdef import OPT_INT
 
 OPT_RECORD = 'typing.Optional[Record]'
